refatoração do about_pages e dashboard_page
