+++ encryption_playground/lib/features/home/presentation/pages/about_page.dart
@@ -1,5 +1,5 @@
-import 'package:encryption_playground/features/home/presentation/pages/about_page_destktop.dart';
-import 'package:encryption_playground/features/home/presentation/pages/about_page_mobile.dart';
+import 'package:encryption_playground/features/home/presentation/pages/about_pages/about_page_destktop.dart';
+import 'package:encryption_playground/features/home/presentation/pages/about_pages/about_page_mobile.dart';
 import 'package:encryption_playground/shared/widgets/responsive_layout_builder.dart';
 import 'package:flutter/material.dart';
