+++ encryption_playground/lib/features/home/presentation/pages/dashboard_page.dart
@@ -1,8 +1,13 @@
+import 'package:encryption_playground/shared/widgets/default_container.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 import '../../../../../shared/theme/app_colors.dart';
 import '../widgets/feature_card.dart';
+
+/*
+* TODO: Place the texts of this page into the internacionalization files
+* */
 
 class DashboardPage extends StatelessWidget {
   final GlobalKey<NavigatorState> contentNavigatorKey;
@@ -17,15 +22,9 @@ class DashboardPage extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           // Header
-          Container(
-            padding: const EdgeInsets.all(32),
-            decoration: BoxDecoration(
-              color: AppColors.darkSurfaceContainer,
-              borderRadius: BorderRadius.circular(2),
-              border: Border.all(color: AppColors.darkOutlineVariant),
-            ),
+          DefaultContainer(
             child: Row(
-              children: [
+            children: [
                 Expanded(
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
@@ -56,11 +55,12 @@ class DashboardPage extends StatelessWidget {
                         label: const Text('Learn more about the project'),
                       ),
                     ],
-                  ),
                 ),
-              ],
-            ),
+              ),
+            ],
+          )
           ),
+
           const SizedBox(height: 48),
 
           Text(
@@ -71,56 +71,64 @@ class DashboardPage extends StatelessWidget {
               color: AppColors.darkOnSurface,
             ),
           ),
+
           const SizedBox(height: 24),
 
           // Feature Cards Grid
-          LayoutBuilder(
-            builder: (context, constraints) {
-              int crossAxisCount = 1;
-              if (constraints.maxWidth > 1200) {
-                crossAxisCount = 4;
-              } else if (constraints.maxWidth > 800) {
-                crossAxisCount = 2;
-              }
-
-              return GridView.count(
-                crossAxisCount: crossAxisCount,
-                shrinkWrap: true,
-                physics: const NeverScrollableScrollPhysics(),
-                mainAxisSpacing: 24,
-                crossAxisSpacing: 24,
-                childAspectRatio: 0.85,
-                children: [
-                  FeatureCard(
-                    icon: Icons.security_outlined,
-                    title: 'Cipher Suite',
-                    description: 'Explore classic and modern substitution ciphers like the Caesar cipher.',
-                    onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/ciphers'),
-                  ),
-                  FeatureCard(
-                    icon: Icons.sync_alt_outlined,
-                    title: 'Symmetric Suite',
-                    description: 'Test symmetric encryption algorithms where the same key encrypts and decrypts.',
-                    onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/symmetric'),
-                  ),
-                  FeatureCard(
-                    icon: Icons.vpn_key_outlined,
-                    title: 'Asymmetric Suite',
-                    description: 'Learn about public-key cryptography and key exchange protocols like Diffie-Hellman.',
-                    onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/asymmetric'),
-                  ),
-                  FeatureCard(
-                    icon: Icons.fingerprint_outlined,
-                    title: 'Hash Suite',
-                    description: 'Generate cryptographic hashes using MD5, SHA-1, SHA-256 and more.',
-                    onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/hash'),
-                  ),
-                ],
-              );
-            },
-          ),
+          _cardsGrid(),
         ],
       ),
     );
   }
+
+  Widget _cardsGrid(){
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        int crossAxisCount = 1;
+        if (constraints.maxWidth > 1200) {
+          crossAxisCount = 4;
+        } else if (constraints.maxWidth > 800) {
+          crossAxisCount = 2;
+        }
+
+        return GridView.count(
+          crossAxisCount: crossAxisCount,
+          shrinkWrap: true,
+          physics: const NeverScrollableScrollPhysics(),
+          mainAxisSpacing: 24,
+          crossAxisSpacing: 24,
+          childAspectRatio: 0.85,
+          children: [
+            FeatureCard(
+              icon: Icons.security_outlined,
+              title: 'Cipher Suite',
+              description: 'Explore classic and modern substitution ciphers like the Caesar cipher.',
+              onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/ciphers'),
+            ),
+            FeatureCard(
+              icon: Icons.sync_alt_outlined,
+              title: 'Symmetric Suite',
+              description: 'Test symmetric encryption algorithms where the same key encrypts and decrypts.',
+              onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/symmetric'),
+            ),
+            FeatureCard(
+              icon: Icons.vpn_key_outlined,
+              title: 'Asymmetric Suite',
+              description: 'Learn about public-key cryptography and key exchange protocols like Diffie-Hellman.',
+              onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/asymmetric'),
+            ),
+            FeatureCard(
+              icon: Icons.fingerprint_outlined,
+              title: 'Hash Suite',
+              description: 'Generate cryptographic hashes using MD5, SHA-1, SHA-256 and more.',
+              onPressed: () => contentNavigatorKey.currentState?.pushReplacementNamed('/hash'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
 }
+
+
